requirements.txt
amqp==5.0.9
asgiref==3.4.1
astroid==2.9.3
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.1
djangorestframework==3.13.1
iniconfig==1.1.1
isort==5.10.1
kombu==5.2.3
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.24
py==1.11.0
pylint==2.12.2
pylint-plugin-utils==0.7
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
